Override the configuration from a well-known file
Chuleisi comes with its own default configuration, and lets admins
provide their own to overwrite parts or all of it.

The application should run properly and safely without any local
configuration file, thanks to the defaults.

Admins can now override this default configuration in
/etc/chuleisi/config.toml which is automatically loaded at startup.

It gets validated to ensure that the configuration is always correct.