config: Introduce the queue configuration
This will allow us to finally process events.

The way it works is deployments will add snippets like the following in
their configuration file:

    [queues.a-name-for-a-queue]
    handler = "some.python.module"

The handler module can be anywhere in Python's sys.path, and the only
requirement right now is that it contains a top-level "process"
function. (the arguments it takes and its return value will eventually
be specified more)

No queues/handlers are configured by default, so the service never pulls
anything out of the box.

Some additional options might be added to the queues in the future.