processor: Process the events
When pulling an event from a queue, this uses the queue configuration
to figure out which handler function to pass the event to, and calls it.

As per previous commits, they needed to be a module containing a
top-level function named "process".

As of now, they are a bit more specified: the process function must take
2 arguments:

* the opened DB session;
* the serialized event as pulled from Redis;