processor: Stop decoding events
There is no reason to assume the events pulled from Redis are UTF-8
strings. They might be just a sequence of bytes corresponding to some
binary format. (e.g a GZip-compressed GVariant serialization)

The event processor handler will know what to do with those bytes, so
leave that responsibility to it.