Add a list of vendors
Vendors often are very creative in what they write on machines as
metadata. As a result the data contains many different names for a given
vendor, which makes queries really hard.

This maps a lot of the… surprising vendor names to their actual names.

We're including it in Azafea rather than as part of an event processor
because it's easy to foresee how that will be useful to many event
processors.