logging: Make SQLAlchemy more verbose
For development, we usually want the logs to contain as much information
as possible. That is the purpose of the verbose configuration option
after all.

This tells SQLAlchemy to print all the SQL queries it runs when Azafea
is running with increased verbosity, which helps a lot understanding
what happens during development.