model: Add a NullableBoolean column type
Booleans can have two values. But sometimes, you want to make it
optional, which adds a third value for None/NULL.

The problem comes when you have a unique constraint on multiple columns,
and one of them is nullable: PotsgreSQL considers all NULL values to be
different.

The NullableBoolean column type helps implement the case where you want
PostgreSQL to consider all the NULL values as identical.

It maps the 3 Python values to an enum of 3 strings:

* True is "true",
* False is "false", and
* None is "unknown";

With that, the unique constraint works as expected.

The only caveat is that the model **MUST** define the column with
`nullable=False`. Otherwise, the column can have a 4th value, NULL,
breaking the unique constraint again. :)