cli: Take responsibility for setting up logging
Now that we have a module dedicated to handling the CLI, it seems like a
much better place to set up the logging based on the configured
verbosity.

It will also allow us to log errors before starting the controller, or
even in parts where we don't have a controller at all, like when
initializing the database.