config: Move the default password checker
Doing it as a config validator meant it was done before setting the
logging.

As a result it did not have the same pretty formatting as everything
else.

In some cases (e.g in some test failures) it would try to log that to a
closed file descriptor, leading to pages of useless tracebacks.