cli: Move back to the __main__ module
In hindsight, this is actually much simpler as it removes the need to
add the source dir to sys.path, and it means the file gets checked by
flake8 and mypy.

The issues I had originally with the module are gone with the
move of most of the code to the cli module.