integration-tests: Clean the model for all tests
Having each test clean its database/model is too much to ask when
writing tests. It is too easy to forget, or to not do it correctly.

And in fact, we weren't doing it thoroughly enough. Specifically, this
wasn't done when a test would fail, causing issues for all subsequent
tests.

This fixes that with a teardown method which is always run, and which
goes a bit further than just cleaning the tables, clearing some
SQLAlchemy internal state.