controller: Exit if all the children have finished
Processors might crash, or they might be killed by the OOM killer, or
they might stop for whatever reason.

When all processors have exited, there is no point in keeping the main
controller around, let's stop it.

This helps admins know that the service has stopped and they need to
investigate what happened and restart it.