integration-tests: Enough with the race conditions
We were starting Azafea in the background, then waiting some time before
killing it, hoping that time had been enough for it to do the work we
wanted to test.

That's inherently racy and fragile: on a less powerful machine, or
depending on the load the machine is under, the tests will randomly
fail.

On the other hand, on a lightly loaded powerful machine, the tests are
waiting longer than needed.

Fortunately, we can now do better with the new exit_on_empty_queues
configuration option. This change uses that new behaviour and removes
all the races. The tests now block for exactly the time needed for
Azafea to start, do its job, and exit.

The only downside is that we have to push the events into Redis before
starting Azafea, but that's not a problem at all in a controled
environment like the tests.