ping v1: Remove the updated_at column
We just talked about it and agreed this didn't make much sense, so let's
remove it before we import the existing data.

In addition, this will make incremental imports much easier, because we
don't need to go over existing configs any more to figure out we need to
update their updated_at.