ping v1: Shave one request off of upserting configs
Currently, when upserting a ping configuration we would:

1. do an `INSERT … ON CONFLICT … UPDATE … RETURNING id` to upsert and
   get the id of the new/existing config;
2. do a `SELECT … FROM … WHERE id = …` to get the ping configuration
   record;
3. use that record for the relationship in the new ping record.

The SELECT query in step 2 is entirely unnecessary as we can directly
pass the config id to the new ping record.

This is what this commit does, so we run one less query for each new
ping record, hopefully speeding things up a bit.