metrics: Add a helper to compute the event date
Dates sent by the client might be wrong for a few reasons: the user
fiddled with the system clock, the system clock has not been
synchronized for some time because the client stayed offline, …

So we have an agorithm to try and correct dates as much as possible.

This requires some calculation based on timestamps in the request and
the event. The request part is cached to avoid computing it again and
again for each event.

Since the formula is not obvious at all, this commit also documents it
by adding here the contents of an internal wiki page which explains the
algorithm.