integration-tests: Better unload handler modules
The code was unloading all the imported modules corresponding to
configured queue handlers.

That was working fine as long as models were defined in the handler
modules.

However, there is no reason event handlers won't split their code into
multiple submodules.

In fact, that's going to happen for the Endless OS metrics.

As a result, for a handler foo.bar.v1 we need to unload not only the
foo.bar.v1 module, but also the foo.bar.model one. To be safe, this
unloads all the modules under foo.bar.