model: Improve the string representation
With sequence events, we are going to have some nullable relationships.

In addition, we're seeing errors in production where some ping events
don't have a ping configuration, and the printing error is masking the
underlying one.

As a result, the method calculating the string representation needs to
take that into account.