metrics: Handle sequence events
With this commit, everything is ready to start handling sequence
events: the registry, the unknown model and the constructor.

The queue processor even processes all the sequence events from the
request.

This is all more complex than singular and aggregate events, because in
the request, we don't get a list of events but a list of sequences, each
sequence containing 2 or more events.

For now we've decided to keep things simple with a few assumptions:

* there is only a start and a stop event, no progress events in between
* the stop event has a null payload

This means we can't handle some of the metrics Endless is collecting,
but we can see about that later, this will already allow us to learn a
lot from the data with simple queries.

No event model were implemented though, so they all end up treated as
unknown.

As an added complexity, the unknown model doesn't store an event but a
whole sequence. Its payload data is then the serialized variant
containing the events in the sequence.