Relicense Azafea to the MPL
Endless are comfortable with keeping public the code handling all the
metrics they collect and process. In fact, they prefer that.

However, I was told Endless can't guarantee that all of their partners
will agree to that, so we unfortunately need to keep a way to have some
(as little as possible) metrics handling code private.

The MPL is a good compromise here, as it enforces copyleft on all the
files licensed as MPL, but allows writing proprietary event processors
should other organizations using Azafea require it.