cli: Allow replaying failed events
We have seen ping events failing in production because of a missing
count field from the incoming records:

    https://github.com/endlessm/azafea/pull/24

That's fixed and we now accept such records. However the failures are
still in the error queue.

This new command helps deal with that, moving all the failures to the
corresponding incoming queue.