metrics: Handle invalid singular events
When an error happens while taking care of an event, the whole request
would normally fail and be pushed back to the error queue in Redis.

However, we would like to instead have only those invalid events to
fail, and the rest of the request to be processed and stored.

This commit does just this, by storing the bad events in their own error
table, much like the unknown events.

Of course, if some other, uncaught failure happens then the whole
request will still fail.