tests: Move some ping checks to unit tests
I had never added unit tests to the event processors because (in
addition to lack of time) I found it weird to have them next to the unit
tests for the Azafea core.

Now that the tests are hosted by the event processors, it becomes
obvious to add unit tests for them.

This actually moves some integration tests to unit tests where it makes
sense, which makes the `test-all` test suite faster and improves the
code coverage of the unit tests.