metrics: Add the possibility to ignore invalid events
Unfortunately, the client code sending us events can have bugs, and that
means it sometimes sends events with invalid payloads.

The most common example is an event requiring a payload, but which is
sent without one.

This adds a list of event UUIDs to ignore when their payload is empty
even though we expected one, so that they don't pile up in the invalid
tables.