config: Better scrub special fields for printing
We currently scrub the passwords when printing the configuration, to
avoid revealing secrets in the logs.

That's done a bit adhoc, replacing the password field in new copies of
the dataclass, before making a dict out of it and serializing it as
TOML.

It turns out we can use our own dict factory, and have it hide all the
passwords.

This not only takes care of hiding all future passwords (provided they
are named "password"), but it also makes it easier to hide other fields
if needed.