cli: Log from the start
An invalid config leads to printing an error before we have setup the
logging, bevause setting up the logging verbosity requires having parsed
the config.

But nothing prevents us from setting up the logging with verbose=False
first, before parsing the config, and then setting it up again with the
configured verbosity.

This is what this commit does.