cli: Prepare for per-queue commands
To allow queue processors to have their own subcommands, we need to know
which queues are configured. But we only know that after parsing the
configuration file, which is provided as a command-line argument.

So we must parse the args once to get the configuration file, parse it,
find the list of configured queues, and then parse the args again.

The core commands are now registered similarly to how per-queue commands
will be.