model: Fix queries on nulleable booleans
The NullableBoolean type allows inserting records with a column which
can be either True, False or None, while properly handling unicity
constraints.

To do that is uses a string enum with three states ('true', 'false' and
'unknown') to the Python values (True, False and None).

However, when trying to filter queries on such a column being None in
Python, SQLAlchemy would generate an SQL with "column IS NULL", instead
or the expected "column = 'unknown'.

This commit fixes that.

Fortunately, we had never done such queries in Python yet, so this bug
didn't actually cause any trouble.