request: Make the serialized column optional
Eventually this column will go away, we're just keeping this for the
first iteration, so that we can reparse it all in case something goes
horribly wrong.

However, it does make the DB bigger.

Since we already have the serialized forms of the historic data
elsewhere (in the Cassandra backups), we decided not to store it for
those old requests we import, and only keep it for the new incoming
requests.

Making the column optional is the first step towards that.