metrics: Ignore the network status changed events
There are **a lot** of them. An order of magniture more than the next
most common event in fact.

And they aren't very useful, because they are too fine-grained, and
require some expensive calculation to query what we're really interested
in: how much time to people spend online vs offline.

So let's remove it and discard it completely, we can add something
better later.