metrics: Allow replaying invalid events
Sometimes a bug will cause a client to send invalid metrics events, or
a bug in Azafea will cause it consider an event invalid, even if it is
valid. When any of those happens, those received events get stored into
the "invalid" tables.

This commit introduces a command to look through the invalid events and
"replay" them as if they had just arrived.

This is useful after fixing bugs in Azafea, or adding invalid events to
the ignore list.