metrics: Allow replaying unknown events
Sometimes Azafea will be a bit late and won't have implemented all the
events the clients send. When that happens, those received events get
stored into the "unknown" tables.

This commit introduces a command to look through the unknown events and
"replay" them as if they had just arrived.

This is useful when we implement new events in Azafea, and want to see
whether we had already received some of them without being able to
process them in the past.