config: Redo the special fields scrubbing
Yes, this is the 3rd attempt.

The problem with the current one is that it can only act on field names,
and we're going to need more than that in subsequent commits.

This new attempt replaces Python's dataclasses.asdict function with a
very similar one, except it filters the returned dicts.