config: Split the queue processor and handler
The only option in the configuration file remains the handler. However a
new, hidden option is introduced: the processor.

The handler now remains the string as configured by the admins, and the
processor is the processing function.

This simplifies things a bit, removing the need for the wrap_with_repr
helper function.

It will also allow automatically discovering more things the way the
'process' function is, most importantly per-queue commands.

The new field should never be set by an admin in the config file though,
and as such is neither documented nor shown when printing the loaded
configuration.