metrics: Explicitly parse the hack clubouse progress payload
Using the generic get_asv_dict here always left me a bit uneasy: it's ok
for a very generic field like the missing codec's "extra_info" which
could contain almost anything, but in this case we do know what we
expect to be in the a{sv} variant.

In addition, there was some miscommunication about the pathways
attribute which turned out to be an array of string variant, rather than
an array of strings as I had expected.

The simple get_asv_dict can't handle that, and so we're getting
NotImplementedError⋅s in production. Oops!

I've toyed with implementing the missing parts in get_asv_dict, but
doing that means basically reimplementing the GLib.Variant.unpack
method, at which point we might as well just use it directly.

In the case of this event though, I do feel we should try to validate
the payload, rather than unpacking it directly.

The main reason for that is that if the unpacking leads to an invalid
dictionary (e.g a required attribute is missing) then we will only catch
it when trying to commit the transaction, which means the whole
transaction will fail.

If we validate here, an invalid payload will lead to the event ending in
the invalid table, while the rest of the request can be processed just
fine.