migrations: Allow running migrations
Queue handlers usually come with their data models. Up until now, the
only way to create them was through the initdb command, which only works
to create new models, but not to modify existing ones.

This commit adds a new `migratedb` command, which will find the
migrations of all configured queues and run them to upgrade the
database.

Relates to #4