migrations: Allow autogenerating migrations
Queue handlers usually come with their data models. Up until now, the
only way to create them was through the initdb command, which only works
to create new models, but not to modify existing ones.

This commit sets up Alembic and adds a new `make-migration` command,
which will try to understand the model for a queue handler and generate
a migration script.

Relates to #4