metrics: Add some PostgreSQL functions
Most queries we want to run on the database are based on the image
"product" and "personality". Having those available as functions makes
this easier.

Creating them as migrations is the easiest way for Azafea to maintain
them.

They take the machine id as argument, and go find the associated image
id in the new Machine mapping table.

While that seems strange (it would have been simpler and more flexible
to have them take directly the image id as their argument), doing it in
this way will allow us to enable row-level security in PostgreSQL: a
given user will only get access to the requests of a machine running a
specific product/personality.