metrics: Allow ignoring replayed unknown aggregate events
In the dev environment, we found some existing aggregate events.

Turns out those are created by the smoke tests of some client libraries.

This means we can finally start implementing and testing the code for
aggregate events!

This commit ignores the aggregate events if they are in the
list of events to ignore when replaying unknown events.