metrics: Avoid useless work when replaying unknown events
We have tens of millions of unknown events in production. The vast
majority are for events we haven't implemented yet.

Trying to replay such a huge quantity of unknown events takes a long
time, and it's completely wasted time because they are still unknown
anyway.

This commit will only replay the events which are now known or ignored,
avoiding useless work.