model: Allow iterating on chunked queries from the end
We use this chunked query to replay events.

Some of the replay commands remove parts of the query as we iterate on
it.

As a result, we try to get chunks until the original end, which
eventually is after the new end, which means the last few chunks are
empty.

This allows these commands to grab chunks starting from the end which
avoids the problem entirely.