metrics: Reverse the chunked queries when replaying unknown/invalid events
We were replaying the events from the start, operating on chunks.

Let's say we had events with ids: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]

Using chunks of size 3, we would expect to get the following subqueries:

1.  slice(0, 3)  => [0, 1, 2]
2.  slice(3, 6)  => [3, 4, 5]
3.  slice(6, 9)  => [6, 7, 8]
4.  slice(9, 10) => [9]

However, the whole point of chunking the query was to avoid taking all
the events in memory at once. As a result the query for each chunk is
evaluated when needed.

And that means it can be evaluated **after** we removed some events from
the global query, for example if they were unknown but aren't any more.

And so what we were really getting was:

1.  slice(0, 3)  => [0, 1, 2]

    let's say the replay removes all 3; now the full query would only
    return [3, 4, 5, 6, 7, 8, 9] !

2.  slice(3, 6)  => [6, 7, 8]

    a whole chunk has been skipped!

    let's say the last 2 are removed; now the full query would only
    return [3, 4, 5, 6, 9] !

3.  slice(6, 9)  => []

    both slice arguments are outside the query. ¯\_(ツ)_/¯

4.  slice(9, 10) => []

    same as the previous step;

As a result, some elements are entirely skipped and never replayed.

The fix is to start with the last chunk and then moving backwards. This
way removals from the query don't impact the next slices.