Enable row-level security
This will allow setting things up on the PostgreSQL side so that most
accounts can only view the user data for the deployments they are
responsible for.

For example, the people at Endless Solutions with access to the database
will only see data for Solutions machines, not for all other users.

The actual security policies will be created and maintained in the
deployment configuration in Terraform, because the azafea user (owning
the database) doesn't have the permissions to do those.