docker: Move to Ubuntu 19.10
We had to work around an issue in GLib's GVariant which has been fixed
in 2.62.1, which is present in this Ubuntu release.

We keep the workaround, at least until the next LTS release of Ubuntu
(which should be out in April 2020).

However, it isn't covered any more by the tests when run in this new
Docker image, so this adds the required pragma to tell the test suite to
ignore the missing coverage of this line.