ci: Better remove the default PostgreSQL
It turns out we weren't removing libpq explicitly, and so APT would
decide that instead of removing it, it would upgrade it and install
postgresql-client-12 to prevent a broken dependency.

This is obviously not what we want, so let's really remove everything
related to the default PostgreSQL, before installing the exact version
we want.