model: Factorize getting the URL
We need two database URLs: one to connect to the DB, and the other to
use in error messages.

For obvious reasons the latter shouldn't contain the password, and as a
result we were building them both independently.

However, SQLAlchemy allows specifying the password (any connection
parameter really, including all those we pass in the URL) separately
from the URL.

Thus we can remove the password from the URL, and use the same for both
the connection and the error messages.