endless: Handle "unknown" image ids
When activations and pings fail to get the image id of a machine, they
report "unknown" because even if we can't know much about the machine,
it is still interesting to count activations/pings in aggregate.

As a result, we can't fail parsing such records. They might be invalid
image ids that don't match the regexp, but they are still valid pings
and activations we must store.

Returning None for each component will allow the event processors to
figure out what to do with them.