metrics: Upsert the machine records
We just made the image id optional for the Machine model so that we can
insert machines when receiving other events than the ImageVersion one.

But if machines can be inserted before we get an image version event,
that means we need to update the image id of existing machines, just as
much as we need to insert new machines with an image id.

Using PostgreSQL's `ON CONFLICT DO UPDATE` construct we can effectively
implement such an "upsert" behaviour.