metrics: Add a dualboot property to the machine
We often want to filter on dualboot machines, and much like the image
this is an immutable property of the machine.

Considering this it makes sense to have it as a column of the Machine
model, false by default and automatically set to true when we receive a
dualboot event for a given machine.

This commit also adds the command to go over old events and set the
machine properties accordingly.