metrics: Add a live property to the machine
We often want to filter on machines running a live USB, and much like
the image this is an immutable property of the machine: rebooting will
change the machine id.

Considering this it makes sense to have it as a column of the Machine
model, false by default and automatically set to true when we receive a
live USB event for a given machine.

This commit also adds the command to go over old events and set the
machine properties accordingly.