metrics: Add a demo property to the machine
We often want to filter on machines running in demo mode, and much like
the image this is an immutable property of the machine: entering or
exiting the demo mode will change the machine id.

Considering this it makes sense to have it as a column of the Machine
model, false by default and automatically set to true when we receive a
demo mode event for a given machine.

We don't need to go over old events here, because those events were
never even processed in the first place until the previous commit.

As a result, what will be needed is to replay unknown events, not to
replay old demo mode events.