metrics: Simplify the deduplication of image version events
Back in 7c29e556137af3b8a1722a330b41f6256971ac84 I fixed a problem with
the image version deduplication: make sure we deduplicate them per
metrics request, rather than globally per DB session.

This was a theoretical issue because in practice, the way Azafea works
we only parse one metrics request per database session.

It was still the right thing to do though.

However, in my eagerness to fix the issue, I found a correct but
wasteful solution. Every time an image version event was added to the
session, the code would:

1. look at all image version in the session
2. group them by request
3. order them by date, multiple times
4. then for each request, keep the oldest image version event

But if we do it each time an image version is added to the session (as
we already were), then we certainly don't need to do it for all
requests, and we know we will always at most have one image version
event older than the one currently being added.

This change simplifies that by only considering the image version events
in the same request as the one being added, and directly expunging that
one if there already was one.

The code is now simpler to understand and it does much less work.