Use the same Dockerfile as we use in production.
In production in Endless, we use a different Dockerfile with a slightly
different setup. The main points are using the Alpine variant of the
Python base image and rendering a `config.toml` file from a template and
environment variable. This commit uses (almost) exactly the same
Dockerfile that we use in production since Azafea was deployed and it
can be argued that it's the most tested one. We don't have a dev/ prod
image so that build argument is dropped.