Improve speed for set-open-durations and remove-os-info-quotes
These commands were retrieving the whole ShellAppIsOpen and OSVersion tables on
the Python backend server, chunk by chunk. This was too slow for the production
server data.

The new version proposed here still uses the ORM, but creates SQL-only requests
that can be launched on the PostgreSQL server. Chunked queries, used to avoid
the load of the whole tables in memory, are:

- useless because the data are not handled by Python, and
- impossible because SQLAlchemy (and I suppose PostgreSQL) cannot edit lines
  using the LIMIT keyword.