Resolve relationships before commit when inserting machine info
When new rows are added, we try to find machine-related information to update
the Machine table accordingly. Unfortunately, as these new rows are not
committed yet, they don’t have a primary key. Without this primary key,
SQLAlchemy doesn’t lazy-load relationships, even if foreign keys are set.

This case is rare in common use cases, as users don’t generally use new ORM
objects’ relationships without committing, or at least flushing the session. As
we want to use this relationship before the commit (we want to insert or update
data in another table, during the same commit as the inserted lines), we
obviously can’t commit before accessing the attribute.

We use enable_relationship_loading to enable relationships lazy-loading, even
when the rows are not inserted yet. The feature is not always enabled in
SQLAlchemy because it has some drawbacks that may lead to inconsistent data in
some cases (backrefs may not automatically be updated for example), but it’s
well adapted to our needs.

Related to #121.