Set short timeout on first Redis pop
The previous code was always triggering a long timeout, even when
executed with the exit_on_empty_queues option. This commit uses a shorter
timeout until the first time the queues are empty.

This is really useful for tests, as most of them are obviously launched with
the exit_on_empty_queues option set: it saves 5 seconds for each test relying
on Redis queues.