Don’t use Redis blocking functions
The previous code was always triggering a long timeout, even when
executed with the exit_on_empty_queues option.

Using sleep() instead of blocking functions gives code that’s more simple, and
that avoids this extra timeout during the first loop.

Another change required by non-blocking functions is that we now use a loop for
different queues. Even if the loop adds 3 extra (trivial) lines, it also
simplifies code, especially for moking Redis.

This is really useful for tests, as most of them are obviously launched with
the exit_on_empty_queues option set: it saves 5 seconds for each test relying
on Redis queues.