metrics: Use uuid5 for v3 monthly aggregate events
Rather than using a random UUID for monthly events that are aggregations
of daily events, use a version 5 name based UUID[1] with the daily event
UUID as the namespace. This is used on the client side to simplify
aggregation as the monthly event UUID can be calculated from the daily
event UUID without any knowledge of what either event represents.

1. https://en.wikipedia.org/wiki/Universally_unique_identifier#Versions_3_and_5_(namespace_name-based)

https://phabricator.endlessm.com/T32670