tests: Mark CLI refresh tests as integration
These tests don't mock postgres and therefore should only be included in
the integration tests. In order to use the decorator, its signature
needs to be added to the local pytest mypy stubs. I suspect that
upstream pytest's stubs now cover this case and the local stubs are no
longer needed, but that's for another day.

https://phabricator.endlessm.com/T32670