tests: Reduce non-integration test coverage target to 65%
A lot of Endless v2/v3 event code only gets covered by the integration
tests, so calculating coverage when skipping those tests misses the
existing 86% mark badly. It would be better to omit some files in this
case such as the cli modules, but pytest-cov doesn't provide an option
to extend the omit setting from the CLI. Maybe it would be better to
just disable coverage in the trimmed test suite, but it's still useful
to see.

https://phabricator.endlessm.com/T32670