Pin dependencies for Read the Docs
Rather than duplicate the Read the Docs requirements in a separate file,
include any needed there in the Pipfile dev-packages and generate the
Read the Docs requirements file from that. This will ensure that Read
the Docs builds the documentation using the same versions of the tools
as our own tests.

This is actually working around an issue with the default dependencies
that Read the Docs will install if they're unspecified. In the previous
configuration, `sphinx` was unspecified in
`requirements.readthedocs.txt` and RTD installs 1.8.5 for this project.
See https://docs.readthedocs.io/en/stable/builds.html#python for
details. That version is now broken with a recent `docutils` update
that's incompatible with `sphinx`. A newer `sphinx` is needed where the
`docutils` version is pinned.

https://phabricator.endlessm.com/T32700