Install system packages for Read the Docs
sphinx needs to import all the modules when building the documentation,
which means it needs the main python requirements to be installed. Some
of those depend on system packages or the wheels can't be compiled by
`pip`. Specify the necessary packages in the RTD `apt_packages` option.

In order to use the `build.apt_packages` option, both `build.os` and
`build.tools` must be specified. The python version is specified to
match the version used in the docker image.

https://phabricator.endlessm.com/T32700