Avoid incrementing channel ID sequence if not needed
When processing a submitted request, we ensure that the request channel
has a corresponding Channel record, creating it if needed.

Previously this was done by attempting to insert a new Channel every
time, and, if this fails due to the uniqueness constraint (on the image
ID and site dictionary) being violated, only then trying to fetch an
existing row.

This is correct in the sense that it will always result in the channel
being present in the table. The problem is that attempting to insert a
channel consumes an id from the underlying sequence that channel IDs are
generated from, and rolling back on the constraint violation does not
roll back the sequence. So, each request consumes a channel ID.

At the time of writing, we have only 681 channels but the highest-value
channel ID is 61 177 182. This is a 32-bit signed integer field.  We
launched this protocol with Endless OS 4 about 1 year ago. So assuming
the number of online Endless OS systems per hour remains constant, we
will exhaust the channel ID sequence in ((2 ** 31) / 61 177 182) - 1 =
~34 years. Not imminent but arguably soon enough to worry about.

Instead, try to retrieve the channel before attempting to insert it if
missing. In case of two processes creating the same channel
simultaneously, we still re-fetch the channel if insert fails.

https://phabricator.endlessm.com/T33969